sidebar refactor
